finally fixed the map error
Nobody breathe, nobody move, I fixed the map error

For the 'name' attribute error from my last commit, I learned that option elements actually don't take name attributes so removing them from the code removed the error.

For the map error, I changed the interface for initialState on RecipeSlice.tsx to match what I had for the JSON interface on Search.tsx. I entered in dummy values in the initialState itself,  then I updated line 9 on Search.tsx where I grab the state from the store. Instead of state.recipe, it's state.recipe.results now to match the interfaces. Last, in the apiCall function, I changed the dispatch from json.results to just json.

So. The user types 'chicken' into the input field and hits search. That calls the api. The data comes back and gets put into the variable json, which is an object. Inside that object is a key called results, which is an array of objects with id, title, image, and imageType inside. That's where the api data goes; it makes lots of objects in that array, each with an id, title, image, and imageType. That data is then dispatched and state is updated.

On line 9, state is grabbed in a variable called 'recipes.' 'useAppSelector' is used to grab the state that is inside state --> recipe --> RESULTS because that's where all the api data went. Then we can loop through that variable 'recipes' and manipulate it.