added more state, api calls
Added two more api calls and state to go with each. Can access data out of information but not out of instruction yet.

Noticed a CSS bug: the hamburger menu doesn't go away when you click outside of it or when you click a link and it takes you to another page.