added pictures, fixed state bug on search
Error page added and styled

Moved interfaces to dedicated files

Styled the search page and fixed a bug.

Bug: The way that I defined state in AllRecipesSlice.tsx, the length will never be 0. So in Search.tsx when I map over the state, I had recipes?.length !== 1, render the recipe cards, or render nothing.
However if a search only came back with one result, nothing would be rendered because the length is 1.

Fixed by making state items optional, removing them from the initialState, and changing the code in Search.tsx to recipes[0].id !== undefined. Might be able to do something like this to help fix the problems with the instructionsSlice