fixed instructions
Organized components a little, fixed instructions.

Realized the data from the api was coming back in an array, not an object, so the way I was going about state wasn't going to work. Added more components to allow me to map over the data and pass it along further and further.

Once fixed, worked on css a little and organized interfaces into one document.